# Assignment 3 (Auth - Login & protect)/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth Login & Protect API",
    "version": "1.0.0",
    "description": "Express API using Supabase Auth for signup, login, logout, and protected resources."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password", "minLength": 6 },
          "metadata": {
            "type": "object",
            "description": "Optional user details stored in Supabase user_metadata.",
            "additionalProperties": true,
            "example": {
              "name": "Alex Morgan",
              "role": "student"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a user account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "201": { "description": "Account created" },
          "400": { "description": "Invalid request or Supabase signup error" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive JWT tokens",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Missing credentials" },
          "401": { "description": "Invalid login credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logout successful" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Read the authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile returned" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Read the authenticated user's dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard returned" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Read public information",
        "responses": {
          "200": { "description": "Public information returned" }
        }
      }
    }
  }
}